requirements.txt
pyotp
pyqrcode
pyserial
pycrypto
